9 feat: VariantService layer included
